requirements.txt
torch
Pillow
transformers
open_clip_torch
pandas
